Added a @protected() decorator so that we could use the following convention: * Private class methods use a single or double underscore * Protected methods use the @protected() decorator
Furthermore there is a settings.py file where the protection
check can be turned on or off by users of this library.
The main reason for this to exist is that checking protection is
a) not always portable between interpreters, and,
b) slow.  Therefore should not be used for functions that get
called a lot and where performance is important.