Added Savable class for saving/loading instance state
The Saveable class defines the concept of an object that can save its
state to a bundle and then create a new instance with the same state
from the bundle. To create the new instance the user can call
the create_from classmethod with the bundle.  This skips the constructor
and creates a new instance directly calling load_instance_state.