Normalised WaitOn
Now a WaitOn has the following meaning:

You can wait on something happening, if while you're waiting it happens,
then you will receive True from wait(), False if it times out and an
Interrupted() exception will be raised if interrupt() was called (and
is supported).

The behaviour of calling wait() again depends on the implementing class.
It may be an 'event' that can happen many time (e.g. a traffic light turns
green) or it may be something that happens once and for all e.g. a point
in time being reached.  In which case wait() would return True immediately
every time after the first (and even the first if the time has been reached).

There is no longer a concept of 'clearing' a wait on to reset it.