Changed abort behaviour and fixed play/pause and RMQ bugs
Now the futrue returned by Process.abort() will have a value True, if the process
is aborted when the task is complete (even if it was not this call that was responsible
for the abort because it was aborted earlier).  Before it was False if the process
was already aborted.

Pause/play in process now work correctly.

RMQ control now works correctly for pause/play/abort.