Fixed default port bug with defaults that evaluate to False
To determine whether a default was specified for a Port the check

	if port.default

was used, which would erroneously miss the cases where a default
was set with a default value that evaluated to False. To solve
this the initial value of a Port's default value is set to the
new variable _NULL which equals an empty tuple. Checking for this
value will tell whether a user specified a default value