Replace get method of Unbundler by key reference
For some as of yet unexplained reason, using the get method a
saved_state, i.e. Unbundler instance, on the BundleKeys.INPUTS key

    saved_state.get(BundleKeys.INPUTS, None)

always resulted in None being returned, even when the key was very
much present in the Bundle. Strangely enough directly accessing
the content through the key, as in:

    saved_state[BundleKeys.INPUTS]

worked just fine. For now we just use the latter method and catching
the KeyError exception