Implement an explicit record of exposed inputs in ProcessSpec
The original concept was to apply expose_inputs call to the
spec as they were defined only ever keeping the resultant state.
However, after discussion it seems that the information that is
lost in this way affects the behavior from the user's perspective
in unexpected ways. Therefore we opt to know keep an explicit
record of the exposed inputs and query that when the user asks
for the exposed_inputs after the fact