New passive implementation of a Persister class
The new Persister implementation is passive in the sense
that it no longer is listening to state transitions of a
Process, but needs to be actively called when it needs to
persist a process. The abstract class defines a save and
a load_checkpoint method. A simple PicklePersister implementation
uses pickles written to disk to store the checkpoints