Properly propagate dynamic_input ports when exposing inputs
The expose inputs functionality was ignoring dynamic input ports
that may have been defined for the spec of the exposed process,
causing any dynamic inputs passed in the namespace of the exposed
process to the parent process to be rejected by the input
validation. We simply remove the check for the dynamic input in
the expose_inputs call and simply expose the dynamic port as well.

Added tests to test this behavior