Remove the concept of DynamicInputPort and DynamicOutputPort
With the new PortNamespaces, it does not make much sense to signify
a dynamic PortNamespaces by populating it with an actual "dynamic"
port. Rather it should just be a property of the namespace itself.
So we remove the DynamicInputPort and DynamicOutputPort and give
the PortNamespace a _is_dynamic attribute which tells whether it
will allow values that do not have an explicit Port.