Support kwargs for PortNamespace constructor in add_port_namespace
If a nested PortNamespace needs to be constructed, the kwargs are
only passed to the constructor of the terminal PortNamespace. Any
intermediate PortNamespaces will use constructor defaults.

On the ProcessSpec level we now also expose the methods
input_namespace and output_namespace to create PortNamespaces in
the input and output port namespaces