Simplify the PortNamespace class
Reduce the port creating and getting methods to just get_port()
This will retrieve a Port in an arbitrary namespace, creating
PortNamespaces in the namespace that do not yet exist. The creation
of the ports is now done in ProcessSpec itself, as PortNamespace
does not provide any methods to directly create Port instances.