Implement absorb() for PortNamespace
The absorb method of PortNamespace allows it to absorb another instance
of PortNamespace to be absorbed into itself, meaning that all attributes,
including the ports, will be copied from the PortNamespace to oneself.
Specific ports can either be excluded or included through the exclude
and include tuple arguments.

This mechanism can now be used by the expose_inputs method of the
ProcessSpec class.