ProcessSpec implement a simple memoryless exposed_inputs method
This is the most basic implementation of the exposed_inputs method for the
ProcessSpec, which allows someone, given an input dictionary, to distill the
sub dictionary that corresponds to the input port namespace of a certain
process class that was exposed, optionally in some namespace. This implementation
is memoryless, which is to say that it does not remember explicitly the exclude
and include rules of the expose call. If the user does not properly namespace
the exposed inputs, name clashes will cause inputs to be returned even though
they were excluded or not included.

The old implementation did maintain an explicit memory and provided the agglomerate
functionality, but was not properly recursive and has been removed for the time
being, until a clean solution can be found