Make PortNamespace.get_port non mutatable and define create_port_namespace
The get_port method could both be used to retrieve an existing Port in
the PortNamespace as well as to create nested PortNamespaces that did
not exist yet. This was ambiguous from the method name. To fix this, the
get_port will now only retrieve ports and raise if any of the intermediate
nested ports does not exist. To create new (nested) namespaces the
create_port_namespace method can be used