Remove obsolete legacy methods from ProcessSpec
The methods dynamic_input, no_dynamic_input and input_validator were
remnants from before the implementation of PortNamespace and were a
way to control the dynamicity of the inputs or ouputs of a process
spec. With the advent of PortNamespaces, the inputs and outputs of
the ProcessSpec are port namespace and these dynamic properties are
attributes of the port namespace. To set them then we should go
directly through the properties of the port namespace