Fix a bug in ProcessSpec._create_port
The create_port method of the ProcessSpec class contained a bug that
was not covered by the tests. When a name was passed that is namespaced
the port name was determined by popping the first element after splitting
the name on the namespace separator. However, the first element would
be the first level of the namespace and not the port name. The actual
port name is the last element of the split namespaced name. Fixed the
bug and added a test that covers this