Ensure that create_input_args respects dynamic input values
The previous commit refactored create_input_args to recursively create
AttributeFrozenDicts for nested port namespaces, but introduced a bug
where dynamic inputs, i.e. values that are not explicitly matched in the
input port namespace, were omitted. The original behavior is restored and
a test was added to cover this