Implement concept of unsuccessful result in FINISHED state
A process needs to be able to return a result that indicates
that the result is unsuccessful. This would then indicated that
in the on_finish method, the outputs should not be checked as
the unsuccessful result already indicates that something went
wrong. If the process result did not already indicate unsuccessfulness
the outputs will be checked and if that validation fails, the
result will be marked unsuccessful after all