Remove the no_wait option from ProcessLauncher._continue
If the no_wait flag was set to True, the _continue method of  the
ProcessLauncher would return True. This signals to RabbitMQ that
the message was successfully completed and therefore it was removed
from the queue. If at that point, before the task actually being
completed, the runner that was executing the task was killed, the
task would be forever lost, as it was not being put back on the
queue. So instead we always return the future from the _continue
method. To still be able to get the pid from the _launch and not
having to wait for the ContinueAction to be completed, we set the
results of the _on_launch_done to the result of the launch future
which holds the pid of the created process