Persist both the raw_inputs and parsed_inputs (#44)
* Persist both the raw_inputs and parsed_inputs

The raw inputs were being persisted and used to recreated the instance
upon loading from a persisted state, however, the parsed inputs were
being recreated everytime from the raw inputs. This means that the
default inputs not explicitly passed were being recreated everytime.
On the plumpy level this is not a problem, but for example in aiida-core
creating a new input node, that is not the exact same as the one created
the first time of the process creation, is problematic. To solve this we
simply persist the parsed inputs as well and reload it from the persisted
state

* Added travis testing file

* Fixed bug if no inputs were supplied

Supplying no inputs would leave the process with a None inputs dictionary
when in fact it should be an empty dictionary (or one with the defaults
populated).