Allow to pass constructor kwargs for PortNamespace.create_port_namespace
These arguments will be used in the creation of the new PortNamespace.
Any sub port namespaces that may have to be created along the way will
not use these arguments but have the defaults. The downside of this
approach can be easily countered by creating the sub spaces in separate
calls, which is preferable to the opposite approach where one would have
to manually reset all the properties to the defaults of the created sub
namespaces.