Keep exposed port memory and overload PortNamespace attributes on absorb (#55)
The exposing of ports of a certain ProcessSpec into a PortNamespace allows
for the definition of exclude or include rules. When the user later wants
to retrieve the exposed ports given a set of inputs, this should be mapped
on the actual exposed ports and not on the ports of the relevant ProcessSpec.
To make this possible, the actual ports that are exposed should be stored
in a memory container, indexed by the Process where the ports came from and
the namespace into which they were exposed.

Additionally, the exposing of ports can be done in an arbitrarily nested
PortNamespace. When exposing a PortNamespace of a given ProcessSpec into
oneself, one should not just copy the ports, but also all the mutable
properties of the donor PortNamespace. This is now properly implemented in
the PortNamespace.absorb() method.