Unified state interruption behaviour
Thus far there are two ways to interrupt a state:
* Pause
* Kill

The behaviour of these can be state dependent however for the default
RUNNING state we do not allow either to interrupt the state immediately
but rather the appropriate action will be taken upon state completion
AND once the next state has been entered.

In the WAITING state either of these will interrupt immediately and the
state will be re-executed on play.  This means that a user implementing
a custom WAITING state should either
1) make sure the execute function is idempotent, or,
2) overwrite the interrupt function such that it does not immediately
interrupt the state, but rather does something like the default RUNNING
state

The Process.execute() function has been changed so that a pause() does
not result in it returning.  Similarly for Process.step(): this function
will block until the termination of the Process.