Implement a `Process.status` attribute that can be set through ProcessActions (#58)
This new functionality is then used by the `PauseAction` process action,
which now takes an optional message that will be set as the status attribute
of the process. Any status that might have already been set is stored, such
that when the process is unpaused, the original status can be restored.