Updated to work with latest kiwipy
The new version of kiwipy has a communicator that can run on a separate
thread.  This way heartbeats should not be missed.  Callbacks are
scheduled on our event loop such that we always deal with things on our
thread and the result are communicated back to the rabbitmq thread via
futures.