Ensure communicator for CommunicatorWrapper is not None
This is enforced by putting an assert on the value of the communicator
in the constructor in the wrapper. This problem was found because in
the `load_instance_state` a `CommunicatorWrapper` was being instantiated
as long as the key `communicator` was present in the `load_context`
regardless of its value. This led to guards in the `Process` class for
the presence of communicator to be misled, because it would hold a
communicator wrapper, but it could not communicate, since its communicator
was None.