Catch TimeoutError of `add_rpc_subscriber` in `Process.init`
The `add_rpc_subscriber` call on the communicator in the `init` call
of a `Process` can timeout with a `TimeoutError`, but this should not
topple the process itself. Instead, it should log the error to notify
the user that the process won't be able to respond to RPC calls. When
it is reloaded, it has another chance to add the RPC subscriber and
continue functioning normally.