Let exception in `load_checkpoint` in `ProcessLauncher._continue` bubble up
If the `load_checkpoint` call excepts, the checkpoint is most likely
corrupt and no other launcher will have a chance of successfully loading
the checkpoint and continuing the process. It makes no sense to reject
the task, which will simply lead to it being requeued by RabbitMQ.
Instead the exception should bubble up and the caller should deal with
it there.