Make the default value for `PortNamespace` an empty dictionary
A `Port` with a default value, nested in a non-required namespace, was
not being generated when creating an instance of the `Process` if no
explicit value was being passed for the parent namespace. The solution
is to simply create an empty dictionary for all port namespaces, even if
no explicit value is passed.