Fix the version requirements for the `pika` and `kiwipy` dependencies
The recently released `pika==1.0.0b2` is the last expected beta and no
more breaking changes should be introduced in the final release. The
`topika` and `kiwipy` libraries have already been updated to be
compatible with this version in `v0.2.0` and `v0.5.0`, respectively.
Since we only depend directly on `kiwipy` and `pika` we update their
version requirements.