Make the `outputs` of a `Process` follow the nesting of its spec
The process specification of processes allows to define an arbitrarily
nested output namespace, however, the final results were mapped onto a
single flat namespace. In addition, the final outputs were not fully
properly validated against the process spec when the process was
finishing, causing for example required output ports that were not
explicitly output to be ignored and the process would still validate.

Here we fix the output port validation and the internal `outputs` stored
for the process will follow the exact same nesting as defined in the
outputs port namespace in the process specification.