Fix bug in process spec validation with default and validator (#106)
The inputs passed to a process were validated with respect to the
process spec, *before* pre processing the inputs to populate ports with
defaults that had not been explicitly defined in the inputs. For a port
with a validator *and* a default, this meant that the validator was
called before the default had been processed and so the validator always
received an empty tuple, instead of the default specified. The solution
is to simply pre-process the inputs before validating them.

By moving the responsibility of input pre-processing from the `Process`
to the `ProcessSpec` the process now just calls the process spec to
validate itself, passing the raw inputs. This in turn then processes
*and* validates the inputs in the same go.