Add support for nested exclude/include rules in `PortnameSpace.absorb`
The `PortNamespace` class supports nested namespace hierarchies by
design. However, the `absorb` method, which allows one to absorb the
ports of one port namespace into another, only allows to absorb the
ports wholesale. There are many use cases where one would want to absorb
the ports, but exclude some nested ports. The `exclude` and `include`
rules did not account for this.

Here we make the `absorb` properly recursive and allow the `exclude` and
`include` rules to also target nested ports. For example if the port `a`
in the namespace `base` needs to be excluded, one can now define
`exclude=('base.a')`.