Fix the implementation of `include` in `PortNamespace.absorb`
The recent introduction of support for nested namespaces in the
`exclude` and `include` rules of the `PortNamespace.absorb` method is
incorrect when `include` rules are specified. The exclude and include
rules need to be properly unwrapped to account for nesting, however,
this needs to happen at different moments. For exclude rules, namespaces
have to be done on the `leaf` level. As soon as the rule matches exactly
the port name that is considered it should be filtered.

With the include rules it works a little different. For rules
representing nested ports, all the namespaces along the way to the leaf
port *have* to be included as well.