Add the concept of a "lazy" `PortNamespace`
The current design will skip the validation of a `PortNamespace` if it
is not required *and* no values are passed for it upon process creation.
However, before validation, the inputs are pre-processed. This entails
that the input dictionary is mapped onto the input port namespace and
any input that is not explicitly specified, but corresponds to a port
with a default, has the field populated with said default value.

This behavior, however, now makes it impossible to skip the validation
of a non-required port namespace by not specifying any inputs, as soon
as it contains any port with a default. Even if no inputs are passed,
the pre-processing will populate the defaults, which will trigger the
validation because the inputs for the namespace are no longer empty.

To make it possible to have non-required namespaces, that contain nested
ports with default values, optionally validated, we introduce a new port
namespace property: `populate_defaults`. If a `PortNamespace` has the
`populate_defaults` set to `False` the pre processing will only be
triggered if there are explicit inputs defined for it. If this is not the
case, the population of defaults is skipped.