Make breadcrumbs integral part of `Port` and `PortNamespaces`
Instead of only passing breadcrumbs through the port namespace during
validation, we make them an integral part of the `Port` class. The
`Port` class has a new property `breadcrumbs` that will be initialized
to a tuple with its own `name` if it is defined, otherwise the tuple
will be empty. The `PortNamespace.__setitem__` method is then updated to
set the breadcrumbs of the `Port` that is inserted within its namespace
to prepend the breadcrumbs of the namespace itself. In this way, the
breadcrumbs of ports will be automatically reflecting their place within
the nested namespace when being added to a namespace.