Add a context to the call of the `Port.validator`
The context will contain the entire input mapping as passed to the top
level namespace at process instantiation. This is useful because often
validators perform validation on a combination of inputs and so in
addition to the value passed to the port of the validator itself, it
will need the inputs of other ports in the namespace as well.

The `Process` constructor passes a shallow copy of the parsed inputs as
the context to the process spec validation. This is because the parsed
inputs should not be altered and if it were passed by reference, the
validators could actually change it.