Allow lambdas for `InputPort` default values
During the creation of the `InputPort` the optional default value will
now be checked if it is a callable and if so evaluated first, before
validating the returned value. In `PortNamespace.pre_process` the same
check needs to be performed when the default value of a port needs to be
used, evaluating it first if the default is a callable. Note that this
is only done for defaults and not for port values themselves, even if
callables are a valid type. If one passes a callable as a value for a
port that accepts them, the user expects to get the callable back after
validation and not the evaluated value.