Do not add empty optional port namespaces to parsed inputs
If a process defines an optional port namespace, the namespace should
only appear in the parsed inputs when explicitly defined in the inputs
of the process. That is to say, if the port namespace is completely
absent from the inputs at construction time, `self.inputs` should not
contain that key, not even an empty dictionary. This way within the
process lifetime one can distinguish if the namespace was defined at
all, even if just an empty dictionary. This was impossible because even
if not specified, an empty dictionary would always be present.