Enable pre-commit and simplify `super` calls (#153)
* Enable pre-commit for continuous integration

Add `pre-commit` step to the `continous-integration` workflow. The
pre-commit hook will run a bunch of standard fixers to ensure
consistency with regards to trailing whitespace, quotes and adding file
encodings to top of the file. In addition, the `prospector` tool is run
which will run `pylint` and we run `yapf` for automated formatting.

* Remove arguments from `super` calls

This is a relic from Python 2 where it is required to pass the class and
instance to `super` calls, but now that we no longer support Python 2 we
can simplify the syntax to simply `super()`.