Fix the `LoopCommunicator.remove_broadcast_subscriber`
The `remove_broadcast_subscriber` method was piping through to the
`remove_task_subscriber` method, which is probably a copy and paste
error that went unnoticed because the entire class was unused and
untested.

Also change the logging level when required attribute of a `Port` is
overridden by a default from `INFO` to `DEBUG` as this is a very common
operation and was way too noisy.