Suppress exception thrown in generator exit in `create_task`
Due to a bug in `contextlib._GeneratorContextManager` in versions of
Python before 3.8, once the context manager in the `run_task` coroutine
of the `futures.create_task` utility the following exception is printed
as soon as the loop is close:

    Exception ignored in: <generator object create_task.<locals>.run_task>
    Traceback (most recent call last):
      File "../plumpy/plumpy/futures.py", line 102, in run_task
      File "/usr/lib/python3.7/contextlib.py", line 158, in __exit__
    AttributeError: 'NoneType' object has no attribute 'exc_info'

This exception is thrown because somehow the `sys` module is already
`None`. The only real fix would be to fix the `contextlib` module but
since this is not feasible, a workaround is introduced here that catches
this specific exception and simply suppresses it.