`Port`: do not call validator if unspecified and port not required
If no value is specified for a non-required port, the validator should
not be called because there is nothing to be validated. The current
behavior of calling the validator regardless, forces all validators to
deal with the `UNSPECIFIED` value that is passed, but this is a global
constant that should remain internal to `plumpy` and not be used by
client code.