Make application of `nest_asyncio` patch explicit (#179)
A new event loop policy is implemented `PlumpyEventLoopPolicy` which
overrides the `get_event_loop` method to return an event loop that is
patched with the `nest_asyncio` library to make it reentrant. It also
makes sure that it always returns the same event loop because once made
reentrant, only a single one is required.

The policy now has to be enabled explicitly by calling the method
`set_event_loop_policy` instead of implicit, as it used to be done in
the constructor of the `Process` class. The event loop policy can be
reset through `reset_event_loop_policy`.

The `new_event_loop` and `set_event_loop` methods provided by `plumpy`
are overridden to raise `NotImplementedError` because this can cause
undefined behavior. Anyway, with the single reentrant event loop policy
that is enforced by `plumpy` creating a new event loop or setting a
different one is unnecessary.

Co-Authored-By: Jason Eu <morty.yu@yahoo.com>