Trigger application of nest patch in `set_event_loop_policy` (#189)
The `PlumpyEventLoopPolicy` is designed to apply the `nest_asyncio`
patch to the event loop to make it reentrant. Once the policy is applied
using `set_event_loop_policy` the idea is that when calling
`asyncio.get_event_loop` the `get_event_loop` of `PlumpyEventLoopPolicy`
is called instead which will apply the patch to the loop if not already
done before.

However, in a Jupyter notebook, `asyncio.get_event_loop` would not
actually call the patched `get_event_loop` of the policy, even do the
policy was set. This was proved by the fact that calling the method
`asyncio.get_event_loop_policy` would return `PlumpyEventLoopPolicy`. It
was not until `asyncio.get_event_loop_policy().get_event_loop()` was
called that the patch policy was called and the loop was patched.

As a workaround the `set_event_loop_policy` is updated to directly after
setting the new policy to call `get_event_loop_policy`. Through an
unknown mechanism, this forces that `asyncio.get_event_loop` will
actually start to call the implementation of the new policy.