use weakref in PROCESS_STACK
A weak reference to the process is used in order to avoid that the
'process stack' context variable itself can keep the process in memory.
Asynchronous function executions scheduled by asyncio's `call_soon`,
`call_later` or `call_at` get individual copies of the context, which
stay in memory as long as the corresponding handler stays in memory
(even if the execution was cancelled).