Fix `UnboundLocalError` in `DefaultObjectLoader.load_object`.
The `load_object` method would except with `UnboundLocalError` if an
identifier was passed and the `importlib.import_module` would raise an
`ImportError` because the exception messages uses the variable `mod`
which was not determined. Instead it should use `mod_name` which is
obtained by splitting the identifier on the colon.

This splitting of the identifier could also raise an exception if it
does not contain a colon and this case was not caught explicitly either.
The `ValueError` is now caught and a more specific error message is
re-raised.

These problems were not discovered because they weren't tested for. The
`test_class_loader.py` file has been renamed to `test_loaders.py`, the
existing tests have been refactored to use `pytest` instead of `unittest`
and new tests have been added for the previously uncovered pathways.