DevOps: Update the `mypy` pre-commit dependency (#246)
The requirement is updated to `mypy==0.930`. This allows us to move the
configuration from `tox.ini` to the `pyproject.toml` file. The config in
`.pre-commit-config.yaml` is also corrected as the include path did not
properly include the `src` starting folder.

The remaining configuration for `tox.ini` is also moved to the
`pyproject.toml` and deleted.