`PortNamespace`: Fix bug in valid type checking of dynamic namespaces (#255)
If a `valid_type` is set on a `PortNamespace` it is automatically made
dynamic. If a namespace is dynamic, it means that it should accept any
nested namespace, however deeply nested. However, the leafs of the
inputs should still respect the `valid_type`. This was not the case
though. As soon as a port namespace was made dynamic, any nested input
namespace would be expected, regardless of the types of the leaf values.

The `PortNamespace.validate_dynamic_ports` is made recursive. If a port
value is a dictionary, it recursively calls itself to validate its
nested values, ultimately making sure that leaf values have a valid
type if one is specified for the namespace.