Workchains: Accept but deprecate conditional predicates returning None
In 800bcf154c0ea0d4576636b95d2ad2285adec266, the behavior of predicates
passed to `_Conditional` instances was changed to raise if anything but
a boolean was returned. This was to catch cases where users would return
non-boolean values by accident, which would anyway would most likely to
broken behavior of the workchain.

However, quite a number of existing implementations used to do something
like the following in a predicate

    def conditional_predicate(self):
        if some_condition is True:
             return True

In the case that `some_condition` was not equal to `True`, the predicate
would return `None` which would be evaluated as "falsy" and so would
function as if `False` had been returned.

In order to not break all implemenations using this behavior, `None` is
still accepted and automatically converted to `False`. A `UserWarning`
is emitted to warn that the behavior is deprecated. This is used in
favor of a `DeprecationWarning` since those are not shown by default,
which means the majority of users wouldn't see the deprecation warning.